Replace `node-sass` with `sass`
`node-sass` has been deprecated for a couple of years and it's always had a ton of issues installing correctly in different node environments.

The docs now recommend using dart-sass instead, which is the actively maintained official sass compiler. This PR swaps out one for the other.

I also ran the included format script. Going to let tests run in CI because I think I'm running into weird behavior with directory changing on Windows which is causing the last two tests to fail.

Hopefully this allows people to continue to use fiddly even on the latest versions of node!